# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API сервиса многостороннего обмена вещами: профили пользователей, вход по JWT\nи объявления о вещах, которые владелец готов обменять.\n\nЗащищённые маршруты читают HttpOnly cookie `access_token` — они помечены замком.\nКнопки «Authorize» здесь нет и не нужно: выполните `POST /auth/login` прямо из этой\nстраницы — браузер сохранит cookie и будет отправлять её со всеми следующими\nзапросами сам. JavaScript до неё не дотянется, поэтому вписать её руками нельзя.",
        "title": "Цепочка обмена — API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/dashboard": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает количество пользователей, ПВЗ, объявлений и обменов с разбивкой по статусам.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin dashboard"
                ],
                "summary": "Статистика для главной страницы админки",
                "responses": {
                    "200": {
                        "description": "Статистика админки",
                        "schema": {
                            "$ref": "#/definitions/dto.DashboardResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.DashboardError"
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "$ref": "#/definitions/dto.DashboardError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/dto.DashboardError"
                        }
                    }
                }
            }
        },
...
        "/auth/login": {
            "post": {
                "description": "Проверяет nickname и password, кладёт JWT на 12 часов в HttpOnly cookie `access_token`. Из Swagger UI cookie сохранится в браузере, дальше защищённые запросы уйдут с ней автоматически.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Вход",
                "parameters": [
                    {
                        "description": "Учётные данные",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Вошедший пользователь и его роль, cookie в заголовке Set-Cookie",
                        "schema": {
                            "$ref": "#/definitions/dto.AuthenticatedUserResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректное тело запроса",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_user_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Неверный nickname или пароль",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_user_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Аккаунт глобально заблокирован",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_user_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_user_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает профиль владельца cookie `access_token`. Удобно для проверки, что вход прошёл.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Текущий пользователь",
                "responses": {
                    "200": {
                        "description": "Профиль и роль текущего пользователя",
                        "schema": {
                            "$ref": "#/definitions/dto.AuthenticatedUserResponse"
                        }
                    },
                    "401": {
                        "description": "Нет или истекла cookie access_token",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_user_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_user_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AuthenticatedUserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deals_broken": {
                    "type": "integer"
                },
                "deals_completed": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "nickname": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.DashboardError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "dto.DashboardResponse": {
            "type": "object",
            "properties": {
                "exchanges": {
                    "$ref": "#/definitions/dto.ExchangeStatisticsResponse"
                },
                "items": {
                    "$ref": "#/definitions/dto.ItemStatisticsResponse"
                },
                "pickup_points_total": {
                    "type": "integer"
                },
                "users_total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ExchangeStatisticsResponse": {
            "type": "object",
            "properties": {
                "cancelled": {
                    "type": "integer"
                },
                "completed": {
                    "type": "integer"
                },
                "confirmed": {
                    "type": "integer"
                },
                "delivered": {
                    "type": "integer"
                },
                "delivering": {
                    "type": "integer"
                },
                "proposed": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ItemStatisticsResponse": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "integer"
                },
                "reserved": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "traded": {
                    "type": "integer"
                },
                "withdrawn": {
                    "type": "integer"
                }
            }
        },
...
        "dto.LoginRequest": {
            "type": "object",
            "properties": {
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "github_com_sweetlife999_chain-of-trades-avito_internal_user_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "HttpOnly cookie, которую ставит POST /auth/login.",
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}